Use mocks for RunnerTest Put createCallbackMock in a trait
