+++ tests/bootstrap.php
@@ -1,0 +1,4 @@
+<?php
+
+require_once('vendor/autoload.php');
+require_once('support/TestHelpers.php');
